% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LVHBA_SVM.R
\name{LVHBA_SVM}
\alias{LVHBA_SVM}
\title{Support Vector Machine Based on LVHBA}
\usage{
LVHBA_SVM(
  features,
  labels,
  T = 3,
  N = 200,
  eta = 0.003,
  alpha = 0.002,
  beta = 0.002,
  gamma1 = 1,
  gamma2 = 1,
  c_0 = 2,
  r = 10,
  ub = 10,
  lb = 10^(-6),
  auto_tuning = FALSE,
  temperature = 0.1
)
}
\arguments{
\item{N}{Total iterations. Default is 200.}

\item{alpha}{Proximal gradient stepsize of \code{x}. Default is 0.002.}

\item{beta}{Proximal gradient stepsize of \code{a} and \code{b}. Default is 0.002.}

\item{auto_tuning}{Whether an auto-hyperparameter-tuning is needed.Default is
\code{FALSE}.}

\item{temperature}{Temperature of simulating annealing method for auto-
hyperparameter-tuning. Default is 0.1.}
}
\value{
\item{x}{}
\item{y}{}
\item{theta}{}
\item{Xconv}{}
\item{Yconv}{}
\item{Thetaconv}{}
\item{Fseq}{}
\item{test_error_rate_seq}{}
\item{Fseq_test}{}
\item{val_error}{}
\item{test_error}{}
\item{test_error_rate}{}
}
\description{
A short description...
}
