% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEHA_LRMC.R
\name{MEHA_LRMC}
\alias{MEHA_LRMC}
\title{Low Rank Matrix Completion Based on MEHA}
\usage{
MEHA_LRMC(
  M_val,
  M_tr,
  M_val_index,
  M_tr_index,
  A,
  B,
  group,
  N = 200,
  alpha = 1e-04,
  beta = 1e-04,
  eta = 1e-04,
  gamma = 10,
  c = 2,
  c_p = 0.48,
  auto_tuning = FALSE,
  temperature = 0.1
)
}
\arguments{
\item{M_val}{to be written...}

\item{M_tr}{to be written...}

\item{M_val_index}{to be written...}

\item{M_tr_index}{to be written...}

\item{A}{to be written...}

\item{B}{to be written...}

\item{group}{Group information.}

\item{N}{Total iterations. Default is 300.}

\item{alpha}{Default is 1e-4.}

\item{beta}{Default is 1e-4.}

\item{eta}{Default is 1e-4.}

\item{gamma}{Default is 10.}

\item{c}{Default is 2}

\item{c_p}{Default is 0.48.}

\item{auto_tuning}{Whether an auto-hyperparameter-tuning is needed.Default is
\code{FALSE}.}

\item{temperature}{Temperature of simulating annealing method for auto-
hyperparameter-tuning. Default is 0.1.}
}
\value{
a list with 7 components:
\describe{
\item{x}{to}
\item{y}{Coefficients}
\item{theta}{to}
\item{Xconv}{The convergence of sequence X}
\item{Yconv}{The convergence of sequence Y}
\item{Thetaconv}{The convergence of sequence theta}
\item{Fseq}{The convergence of upper function}
}
}
\description{
A short description...
}
