% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEHA_LRMC.R
\name{MEHA_LRMC}
\alias{MEHA_LRMC}
\title{Low Rank Matrix Completion Based on MEHA}
\usage{
MEHA_LRMC(
  M_val,
  M_tr,
  M_val_index,
  M_tr_index,
  A,
  B,
  group,
  N = 200,
  alpha = 1e-04,
  beta = 1e-04,
  eta = 1e-04,
  gamma = 10,
  c = 2,
  c_p = 0.48,
  auto_tuning = FALSE,
  temperature = 0.1
)
}
\arguments{
\item{M_val}{Input matrix for validation.}

\item{M_tr}{Input matrix for training.}

\item{M_val_index}{Index of validation entries.}

\item{M_tr_index}{Index of training entries.}

\item{A}{Row feature matrix of n by p.}

\item{B}{Column feature matrix of n by p.}

\item{group}{A vector to describe the feature group information, with each
element representing the specific number of features in each group.}

\item{N}{Total iterations. Default is 300.}

\item{alpha}{Proximal gradient stepsize of \code{x}. Default is 1e-4.}

\item{beta}{Proximal gradient stepsize of \code{a} and \code{b}. Default is 1e-4.}

\item{eta}{Proximal gradient stepsize of the proxima \code{theta}. Default is
1e-4.}

\item{gamma}{Moreau envelope parameter. Default is 10.}

\item{c}{Penalty strength of the Moreau envelope inequality constraint.
Default is 2}

\item{c_p}{Default is 0.48.}

\item{auto_tuning}{Whether an auto-hyperparameter-tuning is needed.Default is
\code{FALSE}.}

\item{temperature}{Temperature of simulating annealing method for auto-
hyperparameter-tuning. Default is 0.1.}
}
\value{
\item{x}{A vector of length (2M+1), where M denotes the total group number.
The first M values are the within-group penalty strengths of row feature
coefficients \code{a}, the second M values are within-group penalty strength
of column feature coefficients \code{b}, the last value is the penalty
strengths of low rank effect matrix \code{Q}.}
\item{a}{Row feature coefficient.}
\item{b}{Column feature coefficient.}
\item{Q}{low rank effect matrix.}
\item{theta_a}{to}
\item{theta_b}{to}
\item{theta_Q}{to}
\item{Xconv}{Describe the relative convergence situation of sequence X,
based on l2-norm.}
\item{Yconv}{Describe the relative convergence situation of sequence Y,
based on l2-norm.}
\item{Thetaconv}{Describe the relative convergence situation of sequence theta,
based on l2-norm.}
\item{Fseq}{The upper function value in each iteration.}
}
\description{
For low rank matrix completion task, suppose that for matrix M of
n by n, we observe some entries and do not have access to the rest. We
denote the row features X of n by p and column features as Z of n by p.
We model the matrix as the sum of a low rank effect \code{Q} and a
linear combination of the row features and the column features. Denote
the coefficients of row features X and column features Z by \code{a} and
\code{b}. $M = X a 1^T + Z b 1^T + Q$.
The purpose of this function is to determine the optimal feature
coefficients (i.e., \code{a} and \code{b}), low rank effect matrix \code{Q}
and the hyperparameters (penalty strength) \code{x} based on the input
training and validation sets using MEHA.
}
