% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEHA_LRMC.R
\name{MEHA_LRMC}
\alias{MEHA_LRMC}
\title{Low Rank Matrix Completion Based on MEHA}
\usage{
MEHA_LRMC(
  M_val,
  M_tr,
  M_val_index,
  M_tr_index,
  A,
  B,
  group,
  N = 200,
  alpha = 1e-04,
  beta = 1e-04,
  eta = 1e-04,
  gamma = 10,
  c = 2,
  c_p = 0.48,
  auto_tuning = FALSE,
  temperature = 0.1
)
}
\arguments{
\item{M_val}{Input matrix for validation with dimensions n by p, with matrix elements equal to 0 at positions other than the validation entries.}

\item{M_tr}{Input matrix for training with dimensions n by p, with matrix elements equal to 0 at positions other than the training entries.}

\item{M_val_index}{Index of validation entries dimensions n by p, which is a binary matrix where the positions of elements equal to 1 represent the coordinates of the validation entries in the matrix.}

\item{M_tr_index}{Index of training entries dimensions n by p, which is a binary matrix where the positions of elements equal to 1 represent the coordinates of the training entries in the matrix.}

\item{A}{Row feature matrix of n by p.}

\item{B}{Column feature matrix of n by p.}

\item{group}{A vector to describe the feature group information, with each
element representing the specific number of features in each group.}

\item{N}{Total iterations. Default is 300.}

\item{alpha}{Proximal gradient stepsize of \code{x}. Default is 1e-4.}

\item{beta}{Proximal gradient stepsize of \code{a} and \code{b}. Default is 1e-4.}

\item{eta}{Proximal gradient stepsize of the proxima \code{theta}. Default is
1e-4.}

\item{gamma}{Moreau envelope parameter. Default is 10.}

\item{c}{\code{\underline{c}} in MEHA, which is used to generate the penalty parameter \code{c_k} in LV-HBA by \code{c_k = \underline{c}(k+1)^p}.
Default is 2}

\item{c_p}{The index in \code{c_k = \underline{c}(k+1)^p}. Default is 0.48.}

\item{auto_tuning}{When alpha, beta, eta are fixed, whether an auto-hyperparameter-tuning is needed.
Default is \code{FALSE}.}

\item{temperature}{Temperature of simulating annealing method for auto-
hyperparameter-tuning. Default is 0.1.}
}
\value{
\item{x}{A vector of length (2M+1), where M denotes the total group number.
The first M values are the within-group penalty strengths of row feature
coefficients \code{a}, the second M values are within-group penalty strength
of column feature coefficients \code{b}, the last value is the penalty
strengths of low rank effect matrix \code{Q}.}
\item{a}{Row feature coefficient.}
\item{b}{Column feature coefficient.}
\item{Q}{low rank effect matrix.}
\item{theta_a}{The moreau envelope parameter in MEHA, which is of the same scale as variable a}
\item{theta_b}{The moreau envelope parameter in MEHA, which is of the same scale as variable b}
\item{theta_Q}{The moreau envelope parameter in MEHA, which is of the same scale as variable Q}
\item{Xconv}{Describe the relative convergence situation of sequence x generated by MEHA,
which is computed by \eqn{||x^{k+1}-x^k|| / ||x^K||} based on l2-norm.}
\item{Yconv}{Describe the relative convergence situation of sequence y generated by MEHA,
which is computed by \eqn{||y^{k+1}-y^k|| / ||y^K||} based on l2-norm.}
\item{Thetaconv}{Describe the relative convergence situation of sequence theta generated by MEHA,
which is computed by \eqn{||theta^{k+1}-theta^k|| / ||theta^K||} based on l2-norm.}
\item{Fseq}{The upper function value sequence generated by the iteration based on validation set.}
}
\description{
For low rank matrix completion task, suppose that for matrix M of
n by n, we observe some entries and do not have access to the rest. We
denote the row features X of n by p and column features as Z of n by p.
We model the matrix as the sum of a low rank effect \code{Q} and a
linear combination of the row features and the column features. Denote
the coefficients of row features X and column features Z by \code{a} and
\code{b}. $M = X a 1^T + Z b 1^T + Q$.
The purpose of this function is to determine the optimal feature
coefficients (i.e., \code{a} and \code{b}), low rank effect matrix \code{Q}
and the hyperparameters (penalty strength) \code{x} based on the input
training and validation sets using MEHA.
}
\references{
Liu, R., Liu, Z., Yao, W., Zeng, S., & Zhang, J. (2024).
"Rethinking Moreau Envelope for Nonconvex Bi-Level Optimization: A Single-loop and Hessian-free Solution Strategy."
Available at: https://openreview.net/forum?id=i6EtCiIK4a

\if{html}{\out{<div class="sourceCode">}}\preformatted{Gao, L., Ye, J. J., Yin, H., Zeng, S., & Zhang, J. (2022).
"Value function based difference-of-convex algorithm for bilevel hyperparameter selection problems."
Available at: https://proceedings.mlr.press/v162/gao22j.html
}\if{html}{\out{</div>}}
}
